// src/app/enums/apiResponse.ts
export interface OpTypeResp {
  data: OpTypeRespData;
  statuscode: number;
  msg: string;
  isVersionValid: boolean;
  isAppValid: boolean;
  checkID: number;
  isPasswordExpired: boolean;
  mobileNo?: any;
  emailID?: any;
  isLookUpFromAPI: boolean;
}

export interface OpTypeRespData {
  assignedOpTypes: AssignedOpType[];
}

interface AssignedOpType {
  serviceID: number;
  name: string;
  isActive: boolean;
}
// ...
export interface NumberListResp {
  data: NumberListRespData;
  statuscode: number;
  msg: string;
  isVersionValid: boolean;
  isAppValid: boolean;
  checkID: number;
  isPasswordExpired: boolean;
  mobileNo?: any;
  emailID?: any;
  isLookUpFromAPI: boolean;
}
// ...
interface NumberListRespData {
  numSeries: NumSery[];
  operators: Operator[];
  cirlces: Cirlce[];
}

interface Cirlce {
  id: number;
  circle: string;
}

interface Operator {
  name: string;
  oid: number;
  operator?: any;
  tollFree: string;
  opType: number;
  isBBPS: boolean;
  isBilling: boolean;
  min: number;
  max: number;
  length: number;
  lengthMax: number;
  startWith: string;
  image: string;
  isPartial: boolean;
  accountName: string;
  accountRemak: string;
  isAccountNumeric: boolean;
  isGroupLeader: boolean;
  commSettingType: number;
  minRange: number;
  maxRange: number;
  rangeId: number;
}

interface NumSery {
  oid: number;
  series: number;
  circleCode: number;
  circleID: number;
  circle?: any;
  number?: any;
}
//#endregion


//#region BalanceResp
export interface PGInitiatePGResponse {
  statuscode: number;
  msg: string;
  status: number;
  pGModelForWeb: PGModelForRedirection;
}
// ...
export interface PGModelForRedirection {
  tid: number,
  pgType: number;
  statuscode: number;
  msg: string;
  url: string;
  keyVals: [][];

  rPayRequest: RazorpayRequest;
  paytmJSRequest: PaytmJSRequest;
}
export interface RazorpayRequest {
  key_id: string,
  order_id: string,
  name: string,
  description: string,
  image: string,
  prefill_name: string,
  prefill_contact: string,
  prefill_email: string,
  callback_url: string,
  cancel_url: string,
  amount: string
}
export interface PaytmJSRequest {
  orderID: string,
  amount: string,
  tokenType: string,
  token: string,
  mid: string
  payMode: string
}
export interface BalanceResp {
  data: BalanceDataResp;
  statuscode: number;
  msg: string;
  isVersionValid: boolean;
  isAppValid: boolean;
  checkID: number;
  isPasswordExpired: boolean;
  mobileNo?: any;
  emailID?: any;
  isLookUpFromAPI: boolean;
  isDTHInfoCall: boolean;
  isShowPDFPlan: boolean;
  sid?: any;
  isOTPRequired: boolean;
  getID: number;
}

interface BalanceDataResp {
  balance: number;
  bCapping: number;
  isBalance: boolean;
  isBalanceFund: boolean;
  uBalance: number;
  uCapping: number;
  isUBalance: boolean;
  isUBalanceFund: boolean;
  bBalance: number;
  bbCapping: number;
  isBBalance: boolean;
  isBBalanceFund: boolean;
  cBalance: number;
  cCapping: number;
  isCBalance: boolean;
  isCBalanceFund: boolean;
  idBalnace: number;
  idCapping: number;
  isIDBalance: boolean;
  isIDBalanceFund: boolean;
  pacakgeBalance: number;
  packageCapping: number;
  isPacakgeBalance: boolean;
  isPacakgeBalanceFund: boolean;
  isP: boolean;
  isPN: boolean;
  commRate: number;
}
// ...
export interface TransectionResp {
  liveID: string;
  transactionID: string;
  statuscode: number;
  msg: string;
  isVersionValid: boolean;
  isAppValid: boolean;
  checkID: number;
  isPasswordExpired: boolean;
  mobileNo?: any;
  emailID?: any;
  isLookUpFromAPI: boolean;
  isDTHInfoCall: boolean;
  isShowPDFPlan: boolean;
  sid?: any;
  isOTPRequired: boolean;
  getID: number;
}